Closes #56: clarify id and identifier (#67)
Summary: Ids are unconstrained strings; updated Id description to include recommendation for CURIE; Dropped Identifiers

This merge implements my recent proposal in #56, which is a slight variation on Alex's proposal.
(It's also basically where VMC was several months ago. :-()
Ideally, we'd have more discussion, but we're under a deadline. So, pressing on.